requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
aiosmtplib==3.0.2
alembic==1.13.2
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
fastapi==0.112.1
flower==2.0.1
frozenlist==1.4.1
h11==0.14.0
humanize==4.10.0
idna==3.7
kombu==5.4.0
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
